requirements.txt
fastapi==0.115.8
jinja2==3.1.5
uvicorn==0.34.0
python-multipart==0.0.20
numpy==1.26.4
opencv-python==4.11.0.86
mediapipe==0.10.20
websockets==14.2
pydantic==2.10.6
pydantic-settings==2.7.1
passlib[bcrpyt]==1.7.4
pyjwt==2.10.1
python-jose[cryptography]
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
sshtunnel==
paramiko==